chore: rename Database to ViewModel
